Extract agent-task PR body rendering
